modify

+++ CommonCore/Milestone0/Libft/libft.h
@@ -27,7 +27,7 @@ typedef struct s_list
 	struct s_list	*next;
 }	t_list;
 
-long		ft_atoi(char *str);
+int		ft_atoi(char *str);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t nmemb, size_t size);
 int		ft_isalnum(int c);
